add CmdbExecutor class
