+++ cmdb/src/main/java/com/alibaba/nacos/cmdb/memory/CmdbProvider.java
@@ -24,8 +24,8 @@ import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.cmdb.core.SwitchAndOptions;
 import com.alibaba.nacos.cmdb.service.CmdbReader;
 import com.alibaba.nacos.cmdb.service.CmdbWriter;
+import com.alibaba.nacos.cmdb.utils.CmdbExecutor;
 import com.alibaba.nacos.cmdb.utils.Loggers;
-import com.alibaba.nacos.cmdb.utils.UtilsAndCommons;
 import com.alibaba.nacos.common.utils.JacksonUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -117,9 +117,9 @@ public class CmdbProvider implements CmdbReader, CmdbWriter {
         initCmdbService();
         load();
         
-        UtilsAndCommons.scheduleCmdbTask(new CmdbDumpTask(), switches.getDumpTaskInterval(), TimeUnit.SECONDS);
-        UtilsAndCommons.scheduleCmdbTask(new CmdbLabelTask(), switches.getLabelTaskInterval(), TimeUnit.SECONDS);
-        UtilsAndCommons.scheduleCmdbTask(new CmdbEventTask(), switches.getEventTaskInterval(), TimeUnit.SECONDS);
+        CmdbExecutor.scheduleCmdbTask(new CmdbDumpTask(), switches.getDumpTaskInterval(), TimeUnit.SECONDS);
+        CmdbExecutor.scheduleCmdbTask(new CmdbLabelTask(), switches.getLabelTaskInterval(), TimeUnit.SECONDS);
+        CmdbExecutor.scheduleCmdbTask(new CmdbEventTask(), switches.getEventTaskInterval(), TimeUnit.SECONDS);
     }
     
     @Override
@@ -203,7 +203,7 @@ public class CmdbProvider implements CmdbReader, CmdbWriter {
             } catch (Exception e) {
                 Loggers.MAIN.error("CMDB-LABEL-TASK {}", "dump failed!", e);
             } finally {
-                UtilsAndCommons.scheduleCmdbTask(this, switches.getLabelTaskInterval(), TimeUnit.SECONDS);
+                CmdbExecutor.scheduleCmdbTask(this, switches.getLabelTaskInterval(), TimeUnit.SECONDS);
             }
         }
     }
@@ -225,7 +225,7 @@ public class CmdbProvider implements CmdbReader, CmdbWriter {
             } catch (Exception e) {
                 Loggers.MAIN.error("DUMP-TASK {}", "dump failed!", e);
             } finally {
-                UtilsAndCommons.scheduleCmdbTask(this, switches.getDumpTaskInterval(), TimeUnit.SECONDS);
+                CmdbExecutor.scheduleCmdbTask(this, switches.getDumpTaskInterval(), TimeUnit.SECONDS);
             }
         }
     }
@@ -269,7 +269,7 @@ public class CmdbProvider implements CmdbReader, CmdbWriter {
             } catch (Exception e) {
                 Loggers.MAIN.error("CMDB-EVENT {}", "event task failed!", e);
             } finally {
-                UtilsAndCommons.scheduleCmdbTask(this, switches.getEventTaskInterval(), TimeUnit.SECONDS);
+                CmdbExecutor.scheduleCmdbTask(this, switches.getEventTaskInterval(), TimeUnit.SECONDS);
             }
         }
     }
