+++ config/src/main/java/com/alibaba/nacos/config/server/monitor/MemoryMonitor.java
@@ -30,18 +30,19 @@ import java.util.concurrent.TimeUnit;
  * @author Nacos
  */
 @Service
-class MemoryMonitor {
+public class MemoryMonitor {
     
     @Autowired
     public MemoryMonitor(AsyncNotifyService notifySingleService) {
         
         ConfigExecutor.scheduleConfigTask(new PrintMemoryTask(), DELAY_SECONDS, DELAY_SECONDS, TimeUnit.SECONDS);
         
-        ConfigExecutor.scheduleConfigTask(new PrintGetConfigResponeTask(), DELAY_SECONDS, DELAY_SECONDS,
-                TimeUnit.SECONDS);
+        ConfigExecutor
+                .scheduleConfigTask(new PrintGetConfigResponeTask(), DELAY_SECONDS, DELAY_SECONDS, TimeUnit.SECONDS);
         
-        ConfigExecutor.scheduleConfigTask(new NotifyTaskQueueMonitorTask(notifySingleService), DELAY_SECONDS,
-                DELAY_SECONDS, TimeUnit.SECONDS);
+        ConfigExecutor
+                .scheduleConfigTask(new NotifyTaskQueueMonitorTask(notifySingleService), DELAY_SECONDS, DELAY_SECONDS,
+                        TimeUnit.SECONDS);
         
     }
